requirements.txt
# pip install <package> && pip freeze | grep <package> >> requirements.txt
pandas==2.2.3
sentence-transformers==3.4.1
tqdm==4.67.1
elasticsearch==8.17.1
openai==1.60.2
pypdf==5.2.0
streamlit==1.41.1
psycopg2-binary==2.9.10
python-dotenv==0.19.0
smolagents
PyPDF2==3.0.1
pgcli==4.1.0
